requirements.txt
beautifulsoup4==4.12.2
Django==5.0.1
numpy==1.26.3
Pillow==10.2.0
python-dotenv==1.0.0
Requests==2.31.0